Move the `misc` module into `uint`
